create zod issue parser
